Move city editor to its own file
